Factor the signup component
